Update sparse tensor conversion from Float to Double

+++ starreco/preprocessing/preprocessor.py
@@ -53,10 +53,10 @@ class Preprocessor:
         indices = np.vstack((coo.row, coo.col))
 
         i = torch.LongTensor(indices)
-        v = torch.FloatTensor(values)
+        v = torch.DoubleTensor(values)
         shape = coo.shape
 
-        return torch.sparse.FloatTensor(i, v, torch.Size(shape))
+        return torch.sparse.DoubleTensor(i, v, torch.Size(shape))
 
     def transform(self, df, return_dataframe = False):
         """
